requirements.txt
flask
flask_sqlalchemy
toml
python-dotenv
pulp